Split gsheets
