+++ frictionless/plugins/gsheets/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import GsheetsPlugin
+from .dialect import GsheetsDialect
+from .parser import GsheetsParser
